+++ src/templates/authors.jsx
@@ -21,7 +21,7 @@ const Separator = styled.div`
   justify-content: center;
   margin: 3em 0;
 
-  @media screen and ${(props) => props.theme.devices.phoneL} {
+  @media screen and ${(props) => props.theme.devices.mobileL} {
     margin: 4em 0;
   }
 
@@ -31,9 +31,9 @@ const Separator = styled.div`
 `;
 
 const Div = styled.div`
-  margin: 2em ${(props) => props.theme.layout.marginPhone};
+  margin: 2em ${(props) => props.theme.layout.marginMobile};
 
-  @media screen and ${(props) => props.theme.devices.phoneL} {
+  @media screen and ${(props) => props.theme.devices.mobileL} {
     margin: 2em ${(props) => props.theme.layout.marginTablet};
   }
 
@@ -47,7 +47,7 @@ const Author = styled.div`
   grid-template-columns: 1fr;
   text-align: center;
 
-  @media screen and ${(props) => props.theme.devices.phoneL} {
+  @media screen and ${(props) => props.theme.devices.mobileL} {
     grid-template-columns: 1fr 1fr 1fr;
     text-align: left;
   }
@@ -69,7 +69,7 @@ const Image = styled(GatsbyImage)`
 const AuthorInfo = styled.div`
   margin: 0;
 
-  @media screen and ${(props) => props.theme.devices.phoneL} {
+  @media screen and ${(props) => props.theme.devices.mobileL} {
     margin: 0 0 0 4rem;
   }
 `;